Refactor ProvisionalAssessmentReportGenerator
Move Stats::ProvisionalAssessmentReportGenerator to Stats::ReportGenerator::ProvisionalAssessment
and then inherit from Stats::ReportGenerator::Base
