+++ app/services/stats/report_generator/provisional_assessment.rb
@@ -1,0 +1,9 @@
+module Stats
+  module ReportGenerator
+    class ProvisionalAssessment < Base
+      def reporter
+        Reports::ProvisionalAssessments
+      end
+    end
+  end
+end
